requirements.txt
google-cloud-logging==1.13.0
iso8601==0.1.12
pyparsing==2.4.2
